requirements.txt
tensorflow>=2.10.0
numpy
matplotlib
seaborn
scikit-learn
Pillow
streamlit
opencv-python
